import/java: extract ImportCloser
