+++ components/blitz/src/ome/formats/importer/cli/CommandLineImporter.java
@@ -932,76 +932,3 @@ public class CommandLineImporter {
     }
 
 }
-
-
-class ImportCloser {
-
-    private final static Logger log = LoggerFactory.getLogger(ImportCloser.class);
-
-    List<ImportProcessPrx> imports;
-    int closed = 0;
-    int errors = 0;
-    int processed = 0;
-
-    ImportCloser(OMEROMetadataStoreClient client) throws Exception {
-        this.imports = getImports(client);
-    }
-
-    void closeCompleted() {
-        for (ImportProcessPrx imPrx : imports) {
-            try {
-                processed++;
-                String logName = imPrx.toString().split("\\s")[0];
-                HandlePrx handle = imPrx.getHandle();
-                if (handle != null) {
-                    Response rsp = handle.getResponse();
-                    if (rsp != null) {
-                        log.info("Done: {}", logName);
-                        imPrx.close();
-                        closed++;
-                        continue;
-                    }
-                }
-                log.info("Running: {}", logName);
-            } catch (Exception e) {
-                errors++;
-                log.warn("Failure accessing service", e);
-            }
-        }
-    }
-
-    int getClosed() {
-        return closed;
-    }
-
-    int getErrors() {
-        return errors;
-    }
-
-    int getProcessed() {
-        return processed;
-    }
-
-    private static List<ImportProcessPrx> getImports(OMEROMetadataStoreClient client) throws Exception {
-        final List<ImportProcessPrx> rv = new ArrayList<ImportProcessPrx>();
-        final ServiceFactoryPrx sf = client.getServiceFactory();
-        final List<String> active = sf.activeServices();
-        for (String service : active) {
-            try {
-                final ServiceInterfacePrx prx = sf.getByName(service);
-                final ImportProcessPrx imPrx = ImportProcessPrxHelper.checkedCast(prx);
-                if (imPrx != null) {
-                    try {
-                        imPrx.ice_ping();
-                        rv.add(imPrx);
-                    } catch (Ice.ObjectNotExistException onee) {
-                        // ignore
-                    }
-                }
-            } catch (Exception e) {
-                log.warn("Failure accessing active service", e);
-            }
-        }
-        return rv;
-    }
-}
